+++ BackgroundSubtractorViBeLBSP.h
@@ -8,7 +8,7 @@
 #define BGSVIBELBSP_USE_COLOR_COMPLEMENT 1
 
 //! defines the default value for BackgroundSubtractorViBe::m_nLBSPThreshold (if needed)
-#define BGSVIBELBSP_DEFAULT_LBSP_ABS_SIMILARITY_THRESHOLD ((uchar)30)
+#define BGSVIBELBSP_DEFAULT_LBSP_ABS_SIMILARITY_THRESHOLD (30)
 //! defines the default value for BackgroundSubtractorViBe::m_fLBSPThreshold (if needed)
 #define BGSVIBELBSP_DEFAULT_LBSP_REL_SIMILARITY_THRESHOLD (0.365f)
 //! defines the default value for BackgroundSubtractorViBeLBSP::m_nColorDistThreshold
@@ -36,17 +36,17 @@
 class BackgroundSubtractorViBeLBSP : public BackgroundSubtractorLBSP {
 public:
 	//! full constructor used to intialize an 'absolute' LBSP-based background subtractor
-	explicit BackgroundSubtractorViBeLBSP(	uchar nLBSPThreshold/*=BGSVIBELBSP_DEFAULT_LBSP_ABS_SIMILARITY_THRESHOLD*/,
-											int nDescDistThreshold=BGSVIBELBSP_DEFAULT_DESC_DIST_THRESHOLD,
-											int nColorDistThreshold=BGSVIBELBSP_DEFAULT_COLOR_DIST_THRESHOLD,
-											int nBGSamples=BGSVIBELBSP_DEFAULT_NB_BG_SAMPLES,
-											int nRequiredBGSamples=BGSVIBELBSP_DEFAULT_REQUIRED_NB_BG_SAMPLES);
+	explicit BackgroundSubtractorViBeLBSP(	size_t nLBSPThreshold/*=BGSVIBELBSP_DEFAULT_LBSP_ABS_SIMILARITY_THRESHOLD*/,
+											size_t nDescDistThreshold=BGSVIBELBSP_DEFAULT_DESC_DIST_THRESHOLD,
+											size_t nColorDistThreshold=BGSVIBELBSP_DEFAULT_COLOR_DIST_THRESHOLD,
+											size_t nBGSamples=BGSVIBELBSP_DEFAULT_NB_BG_SAMPLES,
+											size_t nRequiredBGSamples=BGSVIBELBSP_DEFAULT_REQUIRED_NB_BG_SAMPLES);
 	//! full constructor used to intialize a 'relative' LBSP-based background subtractor
 	explicit BackgroundSubtractorViBeLBSP(	float fLBSPThreshold/*=BGSVIBELBSP_DEFAULT_LBSP_REL_SIMILARITY_THRESHOLD*/,
-											int nDescDistThreshold=BGSVIBELBSP_DEFAULT_DESC_DIST_THRESHOLD,
-											int nColorDistThreshold=BGSVIBELBSP_DEFAULT_COLOR_DIST_THRESHOLD,
-											int nBGSamples=BGSVIBELBSP_DEFAULT_NB_BG_SAMPLES,
-											int nRequiredBGSamples=BGSVIBELBSP_DEFAULT_REQUIRED_NB_BG_SAMPLES);
+											size_t nDescDistThreshold=BGSVIBELBSP_DEFAULT_DESC_DIST_THRESHOLD,
+											size_t nColorDistThreshold=BGSVIBELBSP_DEFAULT_COLOR_DIST_THRESHOLD,
+											size_t nBGSamples=BGSVIBELBSP_DEFAULT_NB_BG_SAMPLES,
+											size_t nRequiredBGSamples=BGSVIBELBSP_DEFAULT_REQUIRED_NB_BG_SAMPLES);
 	//! default destructor
 	virtual ~BackgroundSubtractorViBeLBSP();
 	//! (re)initiaization method; needs to be called before starting background subtraction (note: also reinitializes the keypoints vector)
@@ -58,12 +58,12 @@ public:
 
 protected:
 	//! number of different samples per pixel/block to be taken from input frames to build the background model
-	const int m_nBGSamples;
+	const size_t m_nBGSamples;
 	//! number of similar samples needed to consider the current pixel/block as 'background'
-	const int m_nRequiredBGSamples;
+	const size_t m_nRequiredBGSamples;
 	//! background model pixel intensity samples
 	std::vector<cv::Mat> m_voBGColorSamples;
 	//! absolute per-channel color distance threshold (based on the provided LBSP threshold)
-	const int m_nColorDistThreshold;
+	const size_t m_nColorDistThreshold;
 };
