+++ BackgroundSubtractorViBe_1ch.h
@@ -12,9 +12,9 @@
 class BackgroundSubtractorViBe_1ch : public BackgroundSubtractorViBe {
 public:
 	//! full constructor
-	BackgroundSubtractorViBe_1ch(	int nColorDistThreshold=BGSVIBE_DEFAULT_COLOR_DIST_THRESHOLD,
-									int nBGSamples=BGSVIBE_DEFAULT_NB_BG_SAMPLES,
-									int nRequiredBGSamples=BGSVIBE_DEFAULT_REQUIRED_NB_BG_SAMPLES);
+	BackgroundSubtractorViBe_1ch(	size_t nColorDistThreshold=BGSVIBE_DEFAULT_COLOR_DIST_THRESHOLD,
+									size_t nBGSamples=BGSVIBE_DEFAULT_NB_BG_SAMPLES,
+									size_t nRequiredBGSamples=BGSVIBE_DEFAULT_REQUIRED_NB_BG_SAMPLES);
 	//! default destructor
 	virtual ~BackgroundSubtractorViBe_1ch();
 	//! (re)initiaization method; needs to be called before starting background subtraction
